requirements.txt
fastapi>=0.110.0
python-multipart
uvicorn>=0.28.0
asyncpg>=0.29.0
redis==5.0.3
pyjwt==2.8.0
bcrypt==4.1.2
qrcode[pil]==7.4.2
pydantic-settings==2.2.1
httpx
pytest
pytest-asyncio
